LayoutTests/imported/w3c/web-platform-tests/css/css-text/parsing/text-indent-computed-expected.txt


PASS Property text-indent value '10px'
PASS Property text-indent value '20%'
PASS Property text-indent value 'calc(50% + 60px)'
PASS Property text-indent value '-30px'
PASS Property text-indent value '-40%'
PASS Property text-indent value 'calc(10px - 0.5em)'
FAIL Property text-indent value '10px hanging' assert_equals: expected "10px hanging" but got "10px"
FAIL Property text-indent value '20% each-line' assert_true: '20% each-line' is a supported value for text-indent. expected true got false
FAIL Property text-indent value 'calc(50% + 60px) hanging each-line' assert_true: 'calc(50% + 60px) hanging each-line' is a supported value for text-indent. expected true got false
FAIL Property text-indent value 'each-line hanging calc(10px + 0.5em)' assert_true: 'each-line hanging calc(10px + 0.5em)' is a supported value for text-indent. expected true got false
